Add setup page to support Umbrel situation better; upgrade to .NET 10
