+++ boot.tests/BootRequestIdentityTests.cs
@@ -1,4 +1,5 @@
 using System.Net;
+using boot_portal.Models;
 using boot_portal.Utils;
 using Microsoft.AspNetCore.Http;
