+++ boot.tests/LocalMiningAdapterAuthTests.cs
@@ -1,4 +1,4 @@
-using boot_portal;
+using boot_portal.Models;
 using boot_portal.Services;
 using Microsoft.Extensions.Logging.Abstractions;
