+++ boot.tests/boot.tests.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="MSTest.Sdk/3.6.4">
 
     <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
+        <TargetFramework>net10.0</TargetFramework>
         <LangVersion>latest</LangVersion>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
