+++ boot_portal/Pages/Index.cshtml
@@ -1,4 +1,5 @@
 @page "{addr?}"
+@model boot_portal.Pages.IndexModel
 <!DOCTYPE html>
 <html lang="en">
 <head>
